requirements.txt
psutil
requests 
pyopenssl
cryptography 
python-nmap 
scapy
